fix tests

+++ src/sdk-reference/cpp/1/server/now/snippets/now.cpp
@@ -1,5 +1,5 @@
 try {
-  long long timestamp = kuzzle->server->now();
+  int64_t timestamp = kuzzle->server->now();
 
   std::cout << "Epoch-millis timestamp: " << timestamp << std::endl;
 } catch (kuzzleio::KuzzleException& e) {
