+++ src/sdk-reference/cpp/1/websocket/getters/snippets/getters.cpp
@@ -6,6 +6,6 @@ bool auto_resubscribes = protocol->isAutoResubscribe();
 
 unsigned int port = protocol->getPort();
 
-unsigned long long reconnection_delay = protocol->getReconnectionDelay();
+uint64_t reconnection_delay = protocol->getReconnectionDelay();
 
 bool ssl_connection = protocol->isSslConnection();
